fix dtype

+++ exp/train_torch_tokenmixer-large_on_criteo_kaggle_dataset.py
@@ -126,7 +126,7 @@ NUM_HIDDEN_HEAD = 2
 DIM_HIDDEN_HEAD = 256
 DROPOUT = 0.5
 BIAS = True
-DTYPE = torch.bfloat16
+DTYPE = torch.float32
 
 ####################################################################################################
 #                                  TRAINING SPECIFIC CONFIGURATION                                 #
